move getAuthOptions around
